requirements.txt
numpy==1.14.5
opencv-python==3.4.2.17
Pillow
facenet-pytorch